docs(core): add comprehensive JSDoc annotations and architectural patterns documentation + standartize namings
