+++ __tests__/favoritesStore.test.ts
@@ -1,5 +1,5 @@
 import { expect, test, describe, beforeEach } from 'vitest';
-import { useFavoritesStore } from '../src/store/use_favourites_store';
+import { useFavoritesStore } from '../src/Store/use_favourites_store';
 
 describe('Favorites Store Business Logic', () => {
   beforeEach(() => {
