+++ __tests__/parseQuantity.test.ts
@@ -1,5 +1,5 @@
 import { expect, test, describe } from 'vitest';
-import { parseQuantity } from '../src/lib/utils';
+import { parseQuantity } from '../src/Lib/utils';
 
 describe('Utility: parseQuantity', () => {
   test('should parse standard unit (packs/pcs)', () => {
