+++ src/app/catalog/page.tsx
@@ -2,12 +2,12 @@
 
 import { Suspense } from "react";
 import { useRouter } from "next/navigation";
-import Header from "@/Components/Layout/Header";
-import Footer from "@/Components/Layout/Footer";
-import DealCardFactory from "@/Components/ui/DealCard";
-import { cn } from "@/lib/utils";
-import { CartDrawer } from "@/Components/cart/CartDrawer"; 
-import { useCatalogFacade } from "@/lib/use_catalog_facade";
+import Header from "@/Components/Layout/header";
+import Footer from "@/Components/Layout/footer";
+import DealCardFactory from "@/Components/UI/deal_card";
+import { cn } from "@/Lib/utils";
+import { CartDrawer } from "@/Components/Cart/cart_drawer"; 
+import { useCatalogFacade } from "@/Lib/use_catalog_facade";
 
 function CatalogContent() {
   const router = useRouter();
