--- src/components/ui/SmartImage.tsx
@@ -1,22 +1,0 @@
-"use client";
-
-export default function SmartImage({
-  src,
-  alt,
-  className,
-}: {
-  src: string;
-  alt: string;
-  className?: string;
-}) {
-  return (
-    <div className={`relative h-full w-full overflow-hidden bg-[#2a242a] ${className || ""}`}>
-      <img
-        src={src}
-        alt={alt}
-        loading="lazy"
-        className="absolute inset-0 h-full w-full object-cover transition-transform duration-500 will-change-transform group-hover:scale-[1.05]"
-      />
-    </div>
-  );
-}
