+++ src/lib/use_catalog_strategy.ts
@@ -1,3 +1,10 @@
+/**
+ * @file use_catalog_strategy.ts
+ * @description Defines the data-fetching and categorization strategies for the catalog.
+ * @pattern Strategy Pattern: Encapsulates the specific logic for formatting, categorizing, and retrieving different types of catalog items (Products vs. Recipes) behind a common interface.
+ * @pattern Data Mocking: Uses a helper to multiply smaller datasets into larger arrays to simulate a robust, paginated backend response.
+ */
+
 import { 
   weekDiscounts, 
   dailyDiscounts, 
@@ -7,11 +14,26 @@ import {
   type DealCard 
 } from "@/Data/home_data";
 
+/**
+ * Represents a selectable category/filter tab within a specific catalog strategy.
+ * @property {string} id - The internal identifier used for state matching (e.g., "week-discounts").
+ * @property {string} label - The human-readable name displayed in the UI.
+ * @property {string} slug - The full URL path and query string used for routing.
+ */
+
 export interface CatalogCategory {
   id: string;
   label: string;
   slug: string;
 }
+
+/**
+ * The common interface that all catalog data sources must implement.
+ * Ensures the facade and UI components can interact with any data type interchangeably.
+ * @property {"products" | "recipes"} id - Identifies the strategy domain.
+ * @property {CatalogCategory[]} categories - The list of valid filter categories for this domain.
+ * @property {Function} getData - A method that returns the fully formatted, flattened array of items.
+ */
 
 export interface CatalogStrategy {
   id: "products" | "recipes";
@@ -36,6 +58,11 @@ const formatData = (baseArray: DealCard[], categoryId: string) => {
   }));
 };
 
+/**
+ * Strategy implementation specifically for handling Grocery Products.
+ * Aggregates weekly, daily, and expiring discounts.
+ */
+
 export const ProductStrategy: CatalogStrategy = {
   id: "products",
   categories: [
@@ -51,6 +78,11 @@ export const ProductStrategy: CatalogStrategy = {
   ]
 };
 
+/**
+ * Strategy implementation specifically for handling Culinary Recipes.
+ * Aggregates seasonal picks and community-liked content.
+ */
+
 export const RecipeStrategy: CatalogStrategy = {
   id: "recipes",
   categories: [
@@ -63,6 +95,11 @@ export const RecipeStrategy: CatalogStrategy = {
     ...formatData(peopleLiked || [], "people-liked"),
   ]
 };
+
+/**
+ * A dictionary exporting all available strategies, allowing dynamic lookup 
+ * based on the active tab (e.g., `strategies[activeTab].getData()`).
+ */
 
 export const strategies = {
   products: ProductStrategy,
